Update styling and translations in app
This commit includes multiple adjustments in the app's styling and translations. It removes unused imports of 'getStatusBarHeight' in styles.ts files and changes the margin-top values accordingly. It also includes adjustments in font-size of TextOptions in Questionnaires styles. New translation for the 'Next' button has been added in both English and Portuguese locales. Furthermore, the app's configuration was updated to include new plugins and bundleIdentifier for iOS has been defined.